Updated package.json
The purpose of this boilerplate is to quickly build a react project and be able to deploy right to heroku within the same repository. I moved most of the devDependencies into dependencies in order to allow heroku to build the project and then serve it with the basic express server.